Socket build in Python.
Socket with server-client model. Allowing to create a server and hosting several clients on the same IP.

For creating a server: Specify your server's IPv4 address or use the generic one given in the socket. Change the number of the port according to the goal and the direction of your socket.

For creating a client: It should automatically connect to the server you specified before, in case in doesn't just select the IPv4 wanted and run the script. Change the number of the port according to the goal and the direction of your socket.